control the access in authenticated routes
