+++ src/routes/app.routes.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+
+import Dashboard from '../pages/Dashboard';
+
+const App = createStackNavigator();
+
+const AppRoutes: React.FC = () => (
+  <App.Navigator
+    screenOptions={{
+      headerShown: false,
+      cardStyle: { backgroundColor: '#312e38' },
+    }}
+  >
+    <App.Screen name='Dashboard' component={Dashboard} />
+  </App.Navigator>
+);
+
+export default AppRoutes;
